pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geovideo"
version = "0.1.0"
description = "Generate vertical real-estate map videos from geographic inputs."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "moviepy>=1.0.3",
  "opencv-python>=4.9.0",
  "pillow>=10.2.0",
  "numpy>=1.26.0",
  "pydantic>=2.6.0",
  "typer>=0.9.0",
  "requests>=2.31.0",
]

[project.scripts]
geovideo = "geovideo.cli:app"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["geovideo*"]

[tool.pytest.ini_options]
addopts = "-q"
